feat: add tinygo support
Add a separate build-tagged file for tinygo and not-tinygo (other).

Stub functions not available in tinygo.

Fixes build with:

tinygo build -tags "tinygo purego" -target wasm -no-debug -opt=2 ./
